+++ packages/stream_chat/lib/src/core/http/interceptor/logging_interceptor.dart
@@ -87,8 +87,8 @@ class LoggingInterceptor extends Interceptor {
       requestHeaders['contentType'] = options.contentType?.toString();
       requestHeaders['responseType'] = options.responseType.toString();
       requestHeaders['followRedirects'] = options.followRedirects;
-      requestHeaders['connectTimeout'] = options.connectTimeout;
-      requestHeaders['receiveTimeout'] = options.receiveTimeout;
+      requestHeaders['connectTimeout'] = options.connectTimeout?.toString();
+      requestHeaders['receiveTimeout'] = options.receiveTimeout?.toString();
       _printMapAsTable(_logPrintRequest, requestHeaders, header: 'Headers');
       _printMapAsTable(_logPrintRequest, options.extra, header: 'Extras');
     }
@@ -121,7 +121,7 @@ class LoggingInterceptor extends Interceptor {
   @override
   void onError(DioError err, ErrorInterceptorHandler handler) {
     if (error) {
-      if (err.type == DioErrorType.response) {
+      if (err.type == DioErrorType.badResponse) {
         final uri = err.response?.requestOptions.uri;
         _printBoxed(
           _logPrintError,