+++ packages/stream_chat_flutter/lib/src/utils/extensions.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:diacritic/diacritic.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
@@ -12,6 +14,16 @@ extension StringExtension on String {
   String capitalize() =>
       isNotEmpty ? '${this[0].toUpperCase()}${substring(1).toLowerCase()}' : '';
 
+  /// Returns the biggest line of a text.
+  String biggestLine() {
+    if (contains('\n')) {
+      return split('\n')
+          .reduce((curr, next) => curr.length > next.length ? curr : next);
+    } else {
+      return this;
+    }
+  }
+
   /// Returns whether the string contains only emoji's or not.
   ///
   ///  Emojis guidelines
@@ -22,7 +34,7 @@ extension StringExtension on String {
     if (trimmedString.isEmpty) return false;
     if (trimmedString.characters.length > 3) return false;
     final emojiRegex = RegExp(
-      r'^(\u00a9|\u00ae|[\u2000-\u3300]|\ud83c[\ud000-\udfff]|\ud83d[\ud000-\udfff]|\ud83e[\ud000-\udfff])+$',
+      r'^(\u00a9|\u00ae|\u200d|[\ufe00-\ufe0f]|[\u2600-\u27FF]|[\u2300-\u2bFF]|\ud83c[\ud000-\udfff]|\ud83d[\ud000-\udfff]|\ud83e[\ud000-\udfff])+$',
       multiLine: true,
       caseSensitive: false,
     );
@@ -351,6 +363,33 @@ extension MessageX on Message {
       }
     }
     return copyWith(text: messageTextToRender);
+  }
+
+  /// Returns an approximation of message size
+  double roughMessageSize(double? fontSize) {
+    var messageTextLength = min(text!.biggestLine().length, 65);
+
+    if (quotedMessage != null) {
+      var quotedMessageLength =
+          (min(quotedMessage!.text?.biggestLine().length ?? 0, 65)) + 8;
+
+      if (quotedMessage!.attachments.isNotEmpty) {
+        quotedMessageLength += 8;
+      }
+
+      if (quotedMessageLength > messageTextLength * 1.2) {
+        messageTextLength = quotedMessageLength;
+      }
+    }
+
+    // Quoted message have a smaller font, so it is necessary to reduce the
+    // size of the multiplier to count for the smaller font.
+    var multiplier = 0.55;
+    if (quotedMessage != null) {
+      multiplier = 0.45;
+    }
+
+    return messageTextLength * (fontSize ?? 1) * multiplier;
   }
 
   /// It returns the message with the translated text if available locally