+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,16 +1,114 @@
 ## Upcoming
 
 🐞 Fixed
-[#1463](https://github.com/GetStream/stream-chat-flutter/pull/1463) Fixed support for iPad in the share button for images. 
+
+- [[#1546]](https://github.com/GetStream/stream-chat-flutter/issues/1546)
+  Fixed `StreamMessageInputTheme.linkHighlightColor` returning null for default theme.
+- [[#1548]](https://github.com/GetStream/stream-chat-flutter/issues/1548) Fixed `StreamMessageInput` urlRegex only
+  matching the
+  lowercase `http(s)|ftp`.
+- [[#1542]](https://github.com/GetStream/stream-chat-flutter/issues/1542) Handle error thrown in `StreamMessageInput`
+  when unable to fetch a
+  link preview.
+
+## 6.1.0
+
+🐞 Fixed
+
+- [[#1502]](https://github.com/GetStream/stream-chat-flutter/issues/1502) Fixed `isOnlyEmoji` method Detects Single
+  Hangul
+  Consonants as Emoji.
+- [[#1505]](https://github.com/GetStream/stream-chat-flutter/issues/1505) Fixed Message bubble disappears for Hangul
+  Consonants.
+- [[#1476]](https://github.com/GetStream/stream-chat-flutter/issues/1476) Fixed `UserAvatarTransform.userAvatarBuilder`
+  works only for otherUser.
+- [[#1490]](https://github.com/GetStream/stream-chat-flutter/issues/1490) Fixed `editMessageInputBuilder` property not
+  used in message edit widget.
+- [[#1523]](https://github.com/GetStream/stream-chat-flutter/issues/1523) Fixed `StreamMessageThemeData` not being
+  applied correctly.
+- [[#1525]](https://github.com/GetStream/stream-chat-flutter/issues/1525) Fixed `StreamQuotedMessageWidget` message for
+  deleted messages not being shown correctly.
+- [[#1529]](https://github.com/GetStream/stream-chat-flutter/issues/1529) Fixed `ClipboardData` requires non-nullable
+  string as text on Flutter 3.10.
+- [[#1533]](https://github.com/GetStream/stream-chat-flutter/issues/1533) Fixed `StreamMessageListView` messages grouped
+  incorrectly w.r.t. timestamp.
+- [[#1532]](https://github.com/GetStream/stream-chat-flutter/issues/1532) Fixed `StreamMessageWidget` actions dialog
+  backdrop filter is cut off by safe area.
+
+✅ Added
+
+- Added `MessageTheme.urlAttachmentHostStyle`, `MessageTheme.urlAttachmentTitleStyle`, and
+  `MessageTheme.urlAttachmentTextStyle` to customize the style of the url attachment.
+- Added `StreamMessageInput.ogPreviewFilter` to allow users to filter out the og preview
+  links. [#1338](https://github.com/GetStream/stream-chat-flutter/issues/1338)
+
+  ```dart
+  StreamMessageInput(
+    ogPreviewFilter: (matchedUri, messageText) {
+      final url = matchedUri.toString();
+      if (url.contains('giphy.com')) {
+        // Return false to prevent the OG preview from being built.
+        return false;
+      }
+      // Return true to build the OG preview.
+      return true;
+  ),
+  ```
+
+- Added `StreamMessageInput.hintGetter` to allow users to customize the hint text of the message
+  input. [#1401](https://github.com/GetStream/stream-chat-flutter/issues/1401)
+
+  ```dart
+  StreamMessageInput(
+    hintGetter: (context, hintType) {
+      switch (hintType) {
+        case HintType.searchGif:
+          return 'Custom Search Giphy';
+        case HintType.addACommentOrSend:
+          return 'Custom Add a comment or send';
+        case HintType.slowModeOn:
+          return 'Custom Slow mode is on';
+        case HintType.writeAMessage:
+          return 'Custom Write a message';
+      }
+    },
+  ),
+  ```
+
+- Added `StreamMessageListView.shrinkWrap` to allow users to shrink wrap the message list view.
+
+🔄 Changed
+
+- Updated `dart` sdk environment range to support `3.0.0`.
+- Deprecated `MessageTheme.linkBackgroundColor` in favor of `MessageTheme.urlAttachmentBackgroundColor`.
+- Updated `stream_chat_flutter_core` dependency
+  to [`6.1.0`](https://pub.dev/packages/stream_chat_flutter_core/changelog).
+
+## 6.0.0
+
+🐞 Fixed
+
+- [[#1456]](https://github.com/GetStream/stream-chat-flutter/issues/1456) Fixed logic for showing that a message was
+  read using sending indicator.
+- [[#1462]](https://github.com/GetStream/stream-chat-flutter/issues/1462) Fixed support for iPad in the share button for
+  images.
+- [[#1475]](https://github.com/GetStream/stream-chat-flutter/issues/1475) Fixed typo to fix compilation.
 
 ✅ Added
 
 - Now it is possible to customize the max lines of the title of a url attachment. Before it was always 1 line.
-- Added `attachmentActionsModalBuilder` parameter to `StreamMessageWidget` that allows to customize `AttachmentActionsModal`.
+- Added `attachmentActionsModalBuilder` parameter to `StreamMessageWidget` that allows to
+  customize `AttachmentActionsModal`.
+- Added `StreamMessageInput.sendMessageKeyPredicate` and `StreamMessageInput.clearQuotedMessageKeyPredicate` to
+  customize the keys used to send and clear the quoted message.
 
 🔄 Changed
 
-- Updated `share_plus` dependency to `^6.3.0`
+- Updated dependencies to resolvable versions.
+
+🚀 Improved
+
+- Improved draw of reaction options. [#1455](https://github.com/GetStream/stream-chat-flutter/pull/1455)
 
 ## 5.3.0
 
@@ -19,17 +117,22 @@
 - Updated `photo_manager` dependency to `^2.5.2`
 
 🐞 Fixed
-- [[#1424]](https://github.com/GetStream/stream-chat-flutter/issues/1424) Fixed a render issue when showing messages starting with 4 whitespaces.
+
+- [[#1424]](https://github.com/GetStream/stream-chat-flutter/issues/1424) Fixed a render issue when showing messages
+  starting with 4 whitespaces.
 - Fixed a bug where the `AttachmentPickerBottomSheet` was not able to identify the mobile browser.
 - Fixed uploading files on Windows - fixed temp file path.
 
 ✅ Added
+
 - New `noPhotoOrVideoLabel` displayed when there is no files to choose.
 
 ## 5.2.0
 
 ✅ Added
-- Added a new `bottomRowBuilderWithDefaultWidget` parameter to `StreamMessageWidget` which contains a third parameter (default `BottomRow` widget with `copyWith` method available) to allow easier customization.
+
+- Added a new `bottomRowBuilderWithDefaultWidget` parameter to `StreamMessageWidget` which contains a third parameter (
+  default `BottomRow` widget with `copyWith` method available) to allow easier customization.
 
 🔄 Changed
 
@@ -39,14 +142,20 @@
 - Updated `dart_vlc` dependency to `^0.4.0`
 - Updated `file_picker` dependency to `^5.2.4`
 - Deprecated `StreamMessageWidget.bottomRowBuilder` in favor of `StreamMessageWidget.bottomRowBuilderWithDefaultWidget`.
-- Deprecated `StreamMessageWidget.deletedBottomRowBuilder` in favor of `StreamMessageWidget.bottomRowBuilderWithDefaultWidget`.
+- Deprecated `StreamMessageWidget.deletedBottomRowBuilder` in favor
+  of `StreamMessageWidget.bottomRowBuilderWithDefaultWidget`.
 - Deprecated `StreamMessageWidget.usernameBuilder` in favor of `StreamMessageWidget.bottomRowBuilderWithDefaultWidget`.
 
 🐞 Fixed
-- [[#1379]](https://github.com/GetStream/stream-chat-flutter/issues/1379) Fixed "Issues with photo attachments on web", where the cached image attachment would not render while uploading.
-- Fix render overflow issue with `MessageSearchListTileTitle`. It now uses `Text.rich` instead of `Row`. Better default behaviour and allows `TextOverflow`.
-- [[1346]](https://github.com/GetStream/stream-chat-flutter/issues/1346) Fixed a render issue while uploading video on web.
-- [[#1347]](https://github.com/GetStream/stream-chat-flutter/issues/1347) `onReply` not working in `AttachmentActionsModal` which is used by `StreamImageAttachment` and `StreamImageGroup`.
+
+- [[#1379]](https://github.com/GetStream/stream-chat-flutter/issues/1379) Fixed "Issues with photo attachments on web",
+  where the cached image attachment would not render while uploading.
+- Fix render overflow issue with `MessageSearchListTileTitle`. It now uses `Text.rich` instead of `Row`. Better default
+  behaviour and allows `TextOverflow`.
+- [[1346]](https://github.com/GetStream/stream-chat-flutter/issues/1346) Fixed a render issue while uploading video on
+  web.
+- [[#1347]](https://github.com/GetStream/stream-chat-flutter/issues/1347) `onReply` not working
+  in `AttachmentActionsModal` which is used by `StreamImageAttachment` and `StreamImageGroup`.
 
 ## 5.1.0
 
@@ -80,7 +189,7 @@
 ✅ Added
 
 - Added `StreamMemberGridView` and `StreamMemberListView`.
-- Added support for additional text field params in `StreamMessageInput`
+- Added support for additional text field parameters in `StreamMessageInput`
     * `maxLines`
     * `minLines`
     * `textInputAction`
@@ -265,7 +374,7 @@
 - Fix commands resetting the `StreamMessageInputController.value`.
 - [[#996]](https://github.com/GetStream/stream-chat-flutter/issues/996) Videos break bottom photo
   carousal.
-- Fix: URLs with path and/or query params are not enriched.
+- Fix: URLs with path and/or query parameters are not enriched.
 - [[#1194]](https://github.com/GetStream/stream-chat-flutter/issues/1194) Request permission to
   access gallery when opening the file picker.
 
@@ -929,11 +1038,11 @@ typedef MessageBuilder = Widget Function(
 ## 1.0.0-beta
 
 - **Refreshed widgets design**
-- Improved api documentation
+- Improved API documentation
 - Updated `stream_chat` dependency to `^1.0.0-beta`
-- Extracted sample app into dedicated [repo](https://github.com/GetStream/flutter-samples)
-- Reimplemented existing widgets
-  using [stream_chat_flutter_core](https://pub.dev/packages/stream_chat_flutter_core)
+- Extracted sample app into dedicated [repository](https://github.com/GetStream/flutter-samples)
+- Re-implemented existing widgets
+  using [`stream_chat_flutter_core`](https://pub.dev/packages/stream_chat_flutter_core)
 
 ## 0.2.21
 
@@ -942,7 +1051,7 @@ typedef MessageBuilder = Widget Function(
 
 ## 0.2.20+4
 
-- Fix channelPreview when the message list is empty
+- Fix `channelPreview` when the message list is empty
 
 ## 0.2.20+3