+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 6.1.0
+
+- Updated `dart` sdk environment range to support `3.0.0`.
+- Updated `stream_chat` dependency to [`6.1.0`](https://pub.dev/packages/stream_chat/changelog).
+
+## 6.0.0
+
+- Updated `drift` to `^2.7.0`.
+- Updated dependencies to resolvable versions.
+
 ## 5.1.0
 
 - Reintroduce support for experimental indexedDB on Web.
@@ -18,7 +28,8 @@
 
 ## 4.4.0
 
-- Allowed experimental use of indexedDb on web with `webUseExperimentalIndexedDb` parameter on `StreamChatPersistenceClient`.
+- Allowed experimental use of indexedDb on web with `webUseExperimentalIndexedDb` parameter
+  on `StreamChatPersistenceClient`.
   Thanks [geweald](https://github.com/geweald).
 
 ## 4.3.0